domain: refactor plan replayer implemention (#39084)
* refactor
