+++ domain/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "domainctx.go",
         "optimize_trace.go",
         "plan_replayer.go",
+        "plan_replayer_dump.go",
         "schema_checker.go",
         "schema_validator.go",
         "sysvar_cache.go",
@@ -55,8 +56,10 @@ go_library(
         "//util/logutil",
         "//util/memory",
         "//util/memoryusagealarm",
+        "//util/printer",
         "//util/servermemorylimit",
         "//util/sqlexec",
+        "@com_github_burntsushi_toml//:toml",
         "@com_github_ngaut_pools//:pools",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
